+++ src/main/java/com/ogc/standard/ao/IWeChatAO.java
@@ -8,6 +8,8 @@
  */
 package com.ogc.standard.ao;
 
+import java.math.BigDecimal;
+
 import com.ogc.standard.domain.CallbackResult;
 import com.ogc.standard.dto.res.XN002500Res;
 import com.ogc.standard.dto.res.XN002501Res;
@@ -21,15 +23,15 @@ public interface IWeChatAO {
 
     public XN002500Res getPrepayIdApp(String applyUser, String toUser,
             String payGroup, String refNo, String bizType, String bizNote,
-            Long transAmount, String backUrl);
+            BigDecimal transAmount, String backUrl);
 
     public XN002501Res getPrepayIdH5(String applyUser, String openId,
             String toUser, String payGroup, String refNo, String bizType,
-            String bizNote, Long transAmount, String backUrl);
+            String bizNote, BigDecimal transAmount, String backUrl);
 
     public String getPrepayIdNative(String applyUser, String toUser,
             String payGroup, String refNo, String bizType, String bizNote,
-            Long transAmount, String backUrl);
+            BigDecimal transAmount, String backUrl);
 
     public void doCallbackAPP(String result);
