+++ src/main/java/com/ogc/standard/ao/IWithdrawAO.java
@@ -1,5 +1,6 @@
 package com.ogc.standard.ao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.ogc.standard.annotation.ServiceModule;
@@ -10,28 +11,27 @@ import com.ogc.standard.domain.Withdraw;
 public interface IWithdrawAO {
     String DEFAULT_ORDER_COLUMN = "code";
 
-    //线下取现申请(需交易密码)
-    public String applyOrderTradePwd(String accountNumber, Long amount,
+    // 线下取现申请(需交易密码)
+    public String applyOrderTradePwd(String accountNumber, BigDecimal amount,
             String payCardInfo, String payCardNo, String applyUser,
             String applyNote, String tradePwd);
 
-    //线下取现申请(front/oss 无需交易密码)
-    public String applyOrder(String accountNumber, Long amount,
+    // 线下取现申请(front/oss 无需交易密码)
+    public String applyOrder(String accountNumber, BigDecimal amount,
             String payCardInfo, String payCardNo, String applyUser,
             String applyNote);
 
-    
     public void approveOrder(String code, String approveUser,
             String approveResult, String approveNote, String systemCode);
 
     public void payOrder(String code, String payUser, String payResult,
             String payNote, String channelOrder, String systemCode);
 
-    //分页查询取现订单
+    // 分页查询取现订单
     public Paginable<Withdraw> queryWithdrawPage(int start, int limit,
             Withdraw condition);
 
-    //取现列表查询
+    // 取现列表查询
     public List<Withdraw> queryWithdrawList(Withdraw condition);
 
     public Withdraw getWithdraw(String code, String systemCode);
