+++ src/main/java/com/ogc/standard/api/impl/XN803750.java
@@ -1,5 +1,7 @@
 package com.ogc.standard.api.impl;
 
+import java.math.BigDecimal;
+
 import com.ogc.standard.ao.IWithdrawAO;
 import com.ogc.standard.api.AProcessor;
 import com.ogc.standard.common.JsonUtil;
@@ -25,18 +27,16 @@ public class XN803750 extends AProcessor {
 
     @Override
     public synchronized Object doBusiness() throws BizException {
-        Long amount = StringValidater.toLong(req.getAmount());
+        BigDecimal amount = StringValidater.toBigDecimal(req.getAmount());
         String code = withdrawAO.applyOrderTradePwd(req.getAccountNumber(),
             amount, req.getPayCardInfo(), req.getPayCardNo(),
             req.getApplyUser(), req.getApplyNote(), req.getTradePwd());
         return new PKCodeRes(code);
     }
 
-
-
-	@Override
-	public void doCheck(String inputparams, String operator)
-			throws ParaException {
+    @Override
+    public void doCheck(String inputparams, String operator)
+            throws ParaException {
         req = JsonUtil.json2Bean(inputparams, XN802750Req.class);
         StringValidater.validateBlank(req.getAccountNumber(),
             req.getPayCardInfo(), req.getPayCardNo(), req.getApplyUser(),
