+++ src/main/java/com/ogc/standard/bo/IChargeBO.java
@@ -1,5 +1,6 @@
 package com.ogc.standard.bo;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.ogc.standard.bo.base.IPaginableBO;
@@ -10,13 +11,13 @@ import com.ogc.standard.enums.EJourBizType;
 
 public interface IChargeBO extends IPaginableBO<Charge> {
     String applyOrderOnline(Account account, String payGroup, String refNo,
-            EJourBizType bizType, String bizNote, Long transAmount,
+            EJourBizType bizType, String bizNote, BigDecimal transAmount,
             EChannelType channelType, String applyUser);
 
     void callBackChange(Charge dbCharge, boolean booleanFlag);
 
     String applyOrderOffline(Account account, EJourBizType bizType,
-            Long amount, String payCardInfo, String payCardNo,
+            BigDecimal amount, String payCardInfo, String payCardNo,
             String applyUser, String applyNote);
 
     void payOrder(Charge data, boolean booleanFlag, String payUser,
@@ -27,6 +28,6 @@ public interface IChargeBO extends IPaginableBO<Charge> {
     Charge getCharge(String code, String systemCode);
 
     // 验证渠道今日支付金额是否超限
-    void doCheckTodayPayAmount(String applyUser, Long payAmount,
+    void doCheckTodayPayAmount(String applyUser, BigDecimal payAmount,
             EChannelType channelType, String companyCode, String systemCode);
 }
