+++ src/main/java/com/ogc/standard/bo/IJourHistoryBO.java
@@ -1,5 +1,6 @@
 package com.ogc.standard.bo;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.ogc.standard.bo.base.IPaginableBO;
@@ -14,8 +15,8 @@ import com.ogc.standard.enums.EBoolean;
 public interface IJourHistoryBO extends IPaginableBO<Jour> {
 
     // 对账结果录入
-    public void doCheckJour(Jour jour, EBoolean checkResult, Long checkAmount,
-            String checkUser, String checkNote);
+    public void doCheckJour(Jour jour, EBoolean checkResult,
+            BigDecimal checkAmount, String checkUser, String checkNote);
 
     // 当前流水，调整不通过。即账其实是平的
     public void adjustJourNO(Jour jour, String adjustUser, String adjustNote);
