+++ src/main/java/com/ogc/standard/bo/IWechatBO.java
@@ -8,6 +8,8 @@
  */
 package com.ogc.standard.bo;
 
+import java.math.BigDecimal;
+
 import com.ogc.standard.domain.CompanyChannel;
 import com.ogc.standard.dto.res.XN002500Res;
 import com.ogc.standard.dto.res.XN002501Res;
@@ -19,20 +21,20 @@ import com.ogc.standard.dto.res.XN002501Res;
  */
 public interface IWechatBO {
     public String getPrepayIdApp(CompanyChannel companyChannel, String bizNote,
-            String code, Long transAmount, String ip, String backUrl);
+            String code, BigDecimal transAmount, String ip, String backUrl);
 
     public XN002500Res getPayInfoApp(CompanyChannel companyChannel,
             String payCode, String prepayId);
 
     public String getPrepayIdH5(CompanyChannel companyChannel, String openId,
-            String bizNote, String code, Long transAmount, String ip,
+            String bizNote, String code, BigDecimal transAmount, String ip,
             String bizBackUrl);
 
     public XN002501Res getPayInfoH5(CompanyChannel companyChannel,
             String payCode, String prepayId);
 
     public String getPrepayIdNative(CompanyChannel companyChannel,
-            String bizNote, String code, Long transAmount, String ip,
+            String bizNote, String code, BigDecimal transAmount, String ip,
             String backUrl);
 
 }
